CASM-3589: Fix certain cases in repository creation
This commit changes nexus-repositories-create to ensure that the results
file is only created if at least one repository was successfully
uploaded. This prevents against the product catalog step receiving data
that is potentially malformed.

This commit also handles the case when nexus-setup processed an empty
nexus-repositories.yaml file. In this case, write out an empty
dictionary to the results so that the product catalog does not receive
empty data.

Test Description: Ran the container locally and inspected the results
file.