CASM-3588: Record results of docker upload
This commit adds a function called record_result that will output the
results of the `skopeo sync` command in a format accepted by the product
catalog. This also partially addresses cases in which the upload is
unsuccessful or only partially successful, by ensuring the script only
outputs valid YAML that is also possible to insert into the catalog.

Test Description: Ran the container locally and inspected the resulting
records.yaml file.