Casm 3588 record docker results (#20)
* CASM-3589: Fix a bug in which newlines were mistakenly added to results

This commit changes the way the repo_members array is constructed,
because array members were incorrectly getting newlines added.

Test Description: ran iuf-nexus-setup and inspected resulting file.

* CASM-3589: Fix certain cases in repository creation

This commit changes nexus-repositories-create to ensure that the results
file is only created if at least one repository was successfully
uploaded. This prevents against the product catalog step receiving data
that is potentially malformed.

This commit also handles the case when nexus-setup processed an empty
nexus-repositories.yaml file. In this case, write out an empty
dictionary to the results so that the product catalog does not receive
empty data.

Test Description: Ran the container locally and inspected the results
file.

* CASM-3588: Record results of docker upload

This commit adds a function called record_result that will output the
results of the `skopeo sync` command in a format accepted by the product
catalog. This also partially addresses cases in which the upload is
unsuccessful or only partially successful, by ensuring the script only
outputs valid YAML that is also possible to insert into the catalog.

Test Description: Ran the container locally and inspected the resulting
records.yaml file.

* CASM-3590: Output results of helm upload in iuf-helm-upload

Test Description: I ran the container locally and inspected the
resulting records.yaml file.