CASM-3589: Ensure that string fields are always strings
This commit changes the 'record_result' functions that record component
data in the results files to ensure that string data is always recorded
as a string. Otherwise, this could lead to errors inserting the data
into the catalog, e.g. if a container image version could be parsed as
an integer or float.

This commit also increments the container image version.

Test Description: I ran each of the affected commands and inspected the
resulting YAML by using a YAML parser to ensure it was valid YAML.