CASM-3859: Ensure results file exists (#21)
* CASM-3859: Ensure results file exists

* Add a script called ensure-results-file-exists
* Add 'trap' calls in the following scripts:
  - iuf-helm-upload (nexus-helm-load operation)
  - iuf-nexus-setup (nexus-setup operation)
  - iuf-skopeo-sync (nexus-docker-upload operation)
* Remove now-unnecessary creation of results file in
  nexus-repositories-create, because this is now covered by
  ensure-results-file-exists.

This will fix an error found in IUF where certain cases may not end
up writing out any data to the output file specified using `valueFrom`.
Even though it is still possible to provide a default value in this
case, it gives a misleading error message in the logs. It is probably
best if we do our best to ensure the results file already exists.

Test Description:
* After building the container image locally, I ran each of the 'iuf-' scripts
  locally with empty lists of artifacts to upload and ensured that the file was
  created.

* CASM-3589: Ensure that string fields are always strings

This commit changes the 'record_result' functions that record component
data in the results files to ensure that string data is always recorded
as a string. Otherwise, this could lead to errors inserting the data
into the catalog, e.g. if a container image version could be parsed as
an integer or float.

This commit also increments the container image version.

Test Description: I ran each of the affected commands and inspected the
resulting YAML by using a YAML parser to ensure it was valid YAML.