CASMINST-5920: Handle errors correctly in skopeo and helm (#22)
This commit changes error handling in iuf-skopeo-sync. Prior to this
commit, failures in skopeo as well as errors in the product manifest
(e.g. incorrect docker directory) would result in IUF proceeding on
with the installation. This commit changes the script in two ways:
* Add a nonzero return value to the process() function when the skopeo
  command fails. Giving an invalid docker directory already returns
  a nonzero return value.
* Check the return value of the process() function when processing
  container images in iuf-skopeo-sync, and exit with an exit code of
  1 if any call to process() returned a nonzero return value.

Additionally, this commit changes error handling in iuf-helm-upload.
The iuf-helm-upload script already checks the return code of the
process() function and will exit nonzero if process() encounters
an error. This commit changes the return value of the process()
function when the underlying nexus-upload-repo-helm command fails.

Finally, increment the version by a bugfix version.

Test Description:
* I built the nexus-setup container image locally.
* I ran iuf-skopeo-sync with an invalid docker directory, and checked
  that the process's exit code was nonzero.
* I ran iuf-skopeo-sync with an image which failed to upload, and
  checked that the process's exit code was nonzero.
* I repeated the above with iuf-helm-upload.