CASMINST-5944: Upload paths relative to given docker directory
This commit changes the call to skopeo sync found in the iuf-skopeo-sync
script and the way in which the YAML results are recorded.

Previously, the destination locations for skopeo-sync would be prefixed with
whatever directory was given in the IUF manifest, e.g. if the manifest
specified to use images in the directory
docker/dtr.dev.cray.com/ and an image was found in
docker/dtr.dev.cray.com/path/to/image:TAG then it would be uploaded with
that full path in the name of the image, minus the leading
'docker/' path element, dtr.dev.cray.com/path/to/image:TAG.

If the manifest specified to use images in the directory 'docker/', then
this example image would still be uploaded to the same path,
dtr.dev.cray.com/path/to/image:TAG. Confusingly, if the manifest
specified to use images in the directory 'docker' (without a trailing
slash), then the image would be uploaded with the 'docker/' path element
included, e.g. docker/dtr.dev.cray.com/path/to/image:TAG.

This commit changes the behavior to be more consistent with pre-IUF
behavior as well as what was originally documented in the IUF manifest schema.
As of this commit the script uploads them with paths relative to the
specified directory, e.g. in the example above, if the manifest
specified the directory 'docker', the image would be uploaded as
'dtr.dev.cray.com/path/to/image:TAG', but if the manifest specified the
directory 'docker/dtr.dev.cray.com', then it would be uploaded as
'path/to/image:TAG'.

In a related fix, this commit fixes a bug in the product catalog results
recording, in which the name of the image was assumed to be the basename
of the directory containing a docker manifest.json file. This is
incorrect, as it should similarly be prefixed with leading path elements
(relative to the container image directory specified in the manifest).

Finally, update the version number for the container image.

Test Description:
* I built the image and ran iuf-skopeo-sync locally, inspecting the
  results in Nexus as well as in the results.yaml file.

Co-authored-by: Ryan Haasken <77809410+haasken-hpe@users.noreply.github.com>